Make TupleOp immutable, and cache repr and hash
Each TupleOp gets an empty cache on construction, which is filled on
first call. Objects should not be modified unless by the code that
created (and did not yet share) them. A clone method to get a copy with
empty cache is provided, and the old constructors turned into factory
methods. Hash or value-based comparison should be fast in almost all
cases, due to the caching and other optimisations.
TupleOp (and subclasses) use __slots__ to reduce memory footprint and
speed up attribute access.